requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Bybit API
pybit>=5.6.0

# Technical Analysis
ta>=0.10.2

# LLM Integration
openai>=1.0.0
anthropic>=0.18.0

# Visualization
plotly>=5.18.0

# Testing
pytest>=7.4.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
flask>=2.0.0
